refactor(app): slim robot-admin to restart status
