--- app/src/redux/robot-admin/api-types.ts
@@ -1,5 +1,0 @@
-export interface SystemTimeData {
-  id: 'time'
-  systemTime: string
-  [key: string]: unknown
-}
