+++ app/src/redux/robot-admin/index.ts
@@ -1,3 +1,4 @@
 export * from './actions'
 export * from './constants'
 export * from './selectors'
+export * from './utils'
