+++ app/src/redux/robot-update/constants.ts
@@ -109,3 +109,6 @@ export const ROBOTUPDATE_SET_SESSION_STEP: 'robotUpdate:SET_SESSION_STEP' =
 
 export const ROBOTUPDATE_FILE_UPLOAD_PROGRESS: 'robotUpdate:FILE_UPLOAD_PROGRESS' =
   'robotUpdate:FILE_UPLOAD_PROGRESS'
+
+/** Fallback path when server health does not advertise a restart capability. */
+export const RESTART_PATH: '/server/restart' = '/server/restart'
