+++ app/src/redux/robot-update/epic.ts
@@ -19,7 +19,6 @@ import {
   startDiscovery,
 } from '../discovery/actions'
 import {
-  RESTART_PATH,
   RESTART_PENDING_STATUS,
   RESTART_STATUS_CHANGED,
   RESTART_SUCCEEDED_STATUS,
@@ -49,6 +48,7 @@ import {
   PROCESS_FILE,
   READY_FOR_RESTART,
   RESTART,
+  RESTART_PATH,
   ROBOTUPDATE_CREATE_SESSION,
   ROBOTUPDATE_CREATE_SESSION_SUCCESS,
   ROBOTUPDATE_FILE_INFO,
